Add files via upload
Share Button is very useful in this days as everything is online and everybody uses social media. Share button is used to enable sharing image, audio, link or anything. In the CSS code, you can add different styles to customize the look and feel of the Share button. For example, you can set the background color, font size, border radius, and padding. You can also add hover styles to the button to change its appearance when a user hovers over it. Overall, creating a Share button using HTML, CSS, and JavaScript involves combining the visual styling of the button with the functionality of sharing content, making it easier for users to share your content with others.